pyproject.toml
[tool.poetry]
name = "nfparser"
version = "0.2.0"
description = "NFParser is a tool to automate data extraction from Brazilian electronic invoices (NF-e), including barcode reading, XML download, data extraction, and saving to Excel, streamlining commercial operations."
authors = ["DevDeividMoura <deividcarvalhomoura+github@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
pyzbar = "^0.1.9"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
